Struct plotly::layout::Layout

pub struct Layout { /* fields omitted */ }

Implementations

impl Layout

pub fn new() -> Layout

pub fn title(self, title: Title) -> Layout

pub fn show_legend(self, show_legend: bool) -> Layout

pub fn legend(self, legend: Legend) -> Layout

pub fn margin(self, margin: Margin) -> Layout

pub fn auto_size(self, auto_size: bool) -> Layout

pub fn width(self, width: usize) -> Layout

pub fn height(self, height: usize) -> Layout

pub fn font(self, font: Font) -> Layout

pub fn uniform_text(self, uniform_text: UniformText) -> Layout

pub fn separators(self, separators: &str) -> Layout

pub fn paper_background_color<C: Color>(
self,
paper_background_color: C
) -> Layout

pub fn plot_background_color<C: Color>(self, plot_background_color: C) -> Layout

pub fn color_scale(self, color_scale: LayoutColorScale) -> Layout

pub fn colorway<C: Color>(self, colorway: Vec<C>) -> Layout

pub fn color_axis(self, color_axis: ColorAxis) -> Layout

pub fn mode_bar(self, mode_bar: ModeBar) -> Layout

pub fn hover_mode(self, hover_mode: HoverMode) -> Layout

Determines the mode of hover interactions. If "closest", a single hoverlabel will appear for the "closest" point within the hoverdistance. If "x" (or "y"), multiple hoverlabels will appear for multiple points at the "closest" x- (or y-) coordinate within the hoverdistance, with the caveat that no more than one hoverlabel will appear per trace. If "x unified" (or "y unified"), a single hoverlabel will appear multiple points at the closest x- (or y-) coordinate within the hoverdistance with the caveat that no more than one hoverlabel will appear per trace. In this mode, spikelines are enabled by default perpendicular to the specified axis. If false, hover interactions are disabled. If clickmode includes the "select" flag, hovermode defaults to "closest". If clickmode lacks the "select" flag, it defaults to "x" or "y" (depending on the trace's orientation value) for plots based on cartesian coordinates. For anything else the default value is "closest".

pub fn click_mode(self, click_mode: &str) -> Layout

pub fn drag_mode(self, drag_mode: &str) -> Layout

pub fn select_direction(self, select_direction: &str) -> Layout

pub fn hover_distance(self, hover_distance: i32) -> Layout

pub fn spike_distance(self, spike_distance: i32) -> Layout

pub fn hover_label(self, hover_label: Label) -> Layout

pub fn grid(self, grid: LayoutGrid) -> Layout

pub fn calendar(self, calendar: Calendar) -> Layout

pub fn x_axis(self, xaxis: Axis) -> Layout

pub fn y_axis(self, yaxis: Axis) -> Layout

pub fn x_axis2(self, xaxis: Axis) -> Layout

pub fn y_axis2(self, yaxis: Axis) -> Layout

pub fn x_axis3(self, xaxis: Axis) -> Layout

pub fn y_axis3(self, yaxis: Axis) -> Layout

pub fn x_axis4(self, xaxis: Axis) -> Layout

pub fn y_axis4(self, yaxis: Axis) -> Layout

pub fn x_axis5(self, xaxis: Axis) -> Layout

pub fn y_axis5(self, yaxis: Axis) -> Layout

pub fn x_axis6(self, xaxis: Axis) -> Layout

pub fn y_axis6(self, yaxis: Axis) -> Layout

pub fn x_axis7(self, xaxis: Axis) -> Layout

pub fn y_axis7(self, yaxis: Axis) -> Layout

pub fn x_axis8(self, xaxis: Axis) -> Layout

pub fn y_axis8(self, yaxis: Axis) -> Layout

pub fn annotations(self, annotations: Vec<Annotation>) -> Layout

pub fn add_annotation(&mut self, annotation: Annotation)

pub fn shapes(self, shapes: Vec<Shape>) -> Layout

pub fn add_shape(&mut self, shape: Shape)

pub fn new_shape(self, new_shape: NewShape) -> Layout

pub fn active_shape(self, active_shape: ActiveShape) -> Layout

pub fn template(self, template: &str) -> Layout

pub fn box_mode(self, box_mode: BoxMode) -> Layout

pub fn box_gap(self, box_gap: f64) -> Layout

pub fn box_group_gap(self, box_group_gap: f64) -> Layout

pub fn bar_mode(self, bar_mode: BarMode) -> Layout

pub fn bar_norm(self, bar_norm: BarNorm) -> Layout

pub fn bar_gap(self, bar_gap: f64) -> Layout

pub fn bar_group_gap(self, bar_group_gap: f64) -> Layout

pub fn violin_mode(self, violin_mode: ViolinMode) -> Layout

pub fn violin_gap(self, violin_gap: f64) -> Layout

pub fn violin_group_gap(self, violin_group_gap: f64) -> Layout

pub fn waterfall_mode(self, waterfall_mode: WaterfallMode) -> Layout

pub fn waterfall_gap(self, waterfall_gap: f64) -> Layout

pub fn waterfall_group_gap(self, waterfall_group_gap: f64) -> Layout

pub fn pie_colorway<C: Color>(self, pie_colorway: Vec<C>) -> Layout

pub fn extend_pie_colors(self, extend_pie_colors: bool) -> Layout

pub fn sunburst_colorway<C: Color>(self, sunburst_colorway: Vec<C>) -> Layout

pub fn extend_sunburst_colors(self, extend_sunburst_colors: bool) -> Layout