Struct plotly::scatter::Scatter

pub struct Scatter<X, Y> where
    X: Serialize + Clone + 'static,
    Y: Serialize + Clone + 'static,  { /* fields omitted */ }

Implementations

impl<X, Y> Scatter<X, Y> where
X: Serialize + Clone + 'static,
Y: Serialize + Clone + 'static,

pub fn new<I, K>(x: I, y: K) -> Box<Self> where
I: IntoIterator<Item = X>,
K: IntoIterator<Item = Y>,

pub fn web_gl_mode(self, on: bool) -> Box<Self>

Enables WebGL.

pub fn name(self, name: &str) -> Box<Self>

Sets the trace name. The trace name appear as the legend item and on hover.

pub fn visible(self, visible: Visible) -> Box<Self>

Determines whether or not this trace is visible. If Visible::LegendOnly, the trace is not drawn, but can appear as a legend item (provided that the legend itself is visible).

pub fn show_legend(self, show_legend: bool) -> Box<Self>

Determines whether or not an item corresponding to this trace is shown in the legend.

pub fn legend_group(self, legend_group: &str) -> Box<Self>

Sets the legend group for this trace. Traces part of the same legend group hide/show at the same time when toggling legend items.

pub fn opacity(self, opacity: f64) -> Box<Self>

Sets the opacity of the trace.

pub fn mode(self, mode: Mode) -> Box<Self>

Determines the drawing mode for this scatter trace. If the provided Mode includes "Text" then the text elements appear at the coordinates. Otherwise, the text elements appear on hover. If there are less than 20 points and the trace is not stacked then the default is Mode::LinesMarkers, otherwise it is Mode::Lines.

pub fn ids<S: AsRef<str>>(self, ids: Vec<S>) -> Box<Self>

Assigns id labels to each datum. These ids for object constancy of data points during animation. Should be an array of strings, not numbers or any other type.

pub fn x0<C: NumOrString>(self, x0: C) -> Box<Self>

Alternate to x. Builds a linear space of x coordinates. Use with dx where x0 is the starting coordinate and dx the step.

pub fn dx(self, dx: f64) -> Box<Self>

Sets the x coordinate step. See x0 for more info.

pub fn y0<C: NumOrString>(self, y0: C) -> Box<Self>

Alternate to y. Builds a linear space of y coordinates. Use with dy where y0 is the starting coordinate and dy the step.

pub fn dy(self, dy: f64) -> Box<Self>

Sets the y coordinate step. See y0 for more info.

pub fn text(self, text: &str) -> Box<Self>

Sets text elements associated with each (x,y) pair. If a single string, the same string appears over all the data points. If an array of string, the items are mapped in order to the this trace's (x,y) coordinates. If the trace HoverInfo contains a "text" flag and hover_text is not set, these elements will be seen in the hover labels.

pub fn text_array<S: AsRef<str>>(self, text: Vec<S>) -> Box<Self>

Sets text elements associated with each (x,y) pair. If a single string, the same string appears over all the data points. If an array of string, the items are mapped in order to the this trace's (x,y) coordinates. If trace HoverInfo contains a "text" flag and hover_text is not set, these elements will be seen in the hover labels.

pub fn text_position(self, text_position: Position) -> Box<Self>

Sets the positions of the text elements with respects to the (x,y) coordinates.

pub fn text_position_array(self, text_position: Vec<Position>) -> Box<Self>

Sets the positions of the text elements with respects to the (x,y) coordinates.

pub fn text_template(self, text_template: &str) -> Box<Self>

Template string used for rendering the information text that appear on points. Note that this will override textinfo. Variables are inserted using %{variable}, for example "y: %{y}". Numbers are formatted using d3-format's syntax %{variable:d3-format}, for example "Price: %{y:$.2f}". See format for details on the formatting syntax. Dates are formatted using d3-time-format's syntax %{variable|d3-time-format}, for example "Day: %{2019-01-01|%A}". See format for details on the date formatting syntax. Every attributes that can be specified per-point (the ones that are arrayOk: true) are available.

pub fn text_template_array<S: AsRef<str>>(
self,
text_template: Vec<S>
) -> Box<Self>

Template string used for rendering the information text that appear on points. Note that this will override textinfo. Variables are inserted using %{variable}, for example "y: %{y}". Numbers are formatted using d3-format's syntax %{variable:d3-format}, for example "Price: %{y:$.2f}". See format for details on the formatting syntax. Dates are formatted using d3-time-format's syntax %{variable|d3-time-format}, for example "Day: %{2019-01-01|%A}". See format for details on the date formatting syntax. Every attributes that can be specified per-point (the ones that are arrayOk: true) are available.

pub fn hover_text(self, hover_text: &str) -> Box<Self>

Sets hover text elements associated with each (x,y) pair. If a single string, the same string appears over all the data points. If an array of string, the items are mapped in order to the this trace's (x,y) coordinates. To be seen, trace HoverInfo must contain a "Text" flag.

pub fn hover_text_array<S: AsRef<str>>(self, hover_text: Vec<S>) -> Box<Self>

Sets hover text elements associated with each (x,y) pair. If a single string, the same string appears over all the data points. If an array of string, the items are mapped in order to the this trace's (x,y) coordinates. To be seen, trace HoverInfo must contain a "Text" flag.

pub fn hover_info(self, hover_info: HoverInfo) -> Box<Self>

Determines which trace information appear on hover. If HoverInfo::None or HoverInfo::Skip are set, no information is displayed upon hovering. But, if HoverInfo::None is set, click and hover events are still fired.

pub fn hover_template(self, hover_template: &str) -> Box<Self>

Template string used for rendering the information that appear on hover box. Note that this will override HoverInfo. Variables are inserted using %{variable}, for example "y: %{y}". Numbers are formatted using d3-format's syntax %{variable:d3-format}, for example "Price: %{y:$.2f}". https://github.com/d3/d3-3.x-api-reference/blob/master/Formatting.md#d3_format for details on the formatting syntax. Dates are formatted using d3-time-format's syntax %{variable|d3-time-format}, for example "Day: %{2019-01-01|%A}". https://github.com/d3/d3-3.x-api-reference/blob/master/Time-Formatting.md#format for details on the date formatting syntax. The variables available in hovertemplate are the ones emitted as event data described at this link https://plotly.com/javascript/plotlyjs-events/#event-data. Additionally, every attributes that can be specified per-point (the ones that are arrayOk: true) are available. Anything contained in tag <extra> is displayed in the secondary box, for example "{fullData.name}". To hide the secondary box completely, use an empty tag <extra></extra>.

pub fn hover_template_array<S: AsRef<str>>(
self,
hover_template: Vec<S>
) -> Box<Self>

Template string used for rendering the information that appear on hover box. Note that this will override HoverInfo. Variables are inserted using %{variable}, for example "y: %{y}". Numbers are formatted using d3-format's syntax %{variable:d3-format}, for example "Price: %{y:$.2f}". https://github.com/d3/d3-3.x-api-reference/blob/master/Formatting.md#d3_format for details on the formatting syntax. Dates are formatted using d3-time-format's syntax %{variable|d3-time-format}, for example "Day: %{2019-01-01|%A}". https://github.com/d3/d3-3.x-api-reference/blob/master/Time-Formatting.md#format for details on the date formatting syntax. The variables available in hovertemplate are the ones emitted as event data described at this link https://plotly.com/javascript/plotlyjs-events/#event-data. Additionally, every attributes that can be specified per-point (the ones that are arrayOk: true) are available. Anything contained in tag <extra> is displayed in the secondary box, for example "{fullData.name}". To hide the secondary box completely, use an empty tag <extra></extra>.

pub fn meta<C: NumOrString>(self, meta: C) -> Box<Self>

Assigns extra meta information associated with this trace that can be used in various text attributes. Attributes such as trace name, graph, axis and colorbar title.text, annotation text rangeselector, updatemenues and sliders label text all support meta. To access the trace meta values in an attribute in the same trace, simply use %{meta[i]} where i is the index or key of the meta item in question. To access trace meta in layout attributes, use %{data[n[.meta[i]} where i is the index or key of the meta and n is the trace index.

pub fn custom_data<C: NumOrString>(self, custom_data: Vec<C>) -> Box<Self>

Assigns extra data each datum. This may be useful when listening to hover, click and selection events. Note that, "scatter" traces also appends customdata items in the markers DOM elements

pub fn x_axis(self, axis: &str) -> Box<Self>

Sets a reference between this trace's x coordinates and a 2D cartesian x axis. If "x" ( the default value), the x coordinates refer to Layout::x_axis. If "x2", the x coordinates refer to Layout::x_axis2, and so on.

pub fn y_axis(self, axis: &str) -> Box<Self>

Sets a reference between this trace's y coordinates and a 2D cartesian y axis. If "y" (the default value), the y coordinates refer to Layout::y_axis. If "y2", the y coordinates refer to Layout::y_axis2, and so on.

pub fn orientation(self, orientation: Orientation) -> Box<Self>

Only relevant when stackgroup is used, and only the first orientation found in the stackgroup will be used - including if visible is "legendonly" but not if it is false. Sets the stacking direction. With "v" ("h"), the y (x) values of subsequent traces are added. Also affects the default value of fill.

pub fn group_norm(self, group_norm: GroupNorm) -> Box<Self>

Only relevant when stackgroup is used, and only the first groupnorm found in the stackgroup will be used - including if visible is "legendonly" but not if it is false. Sets the normalization for the sum of this stackgroup. With "fraction", the value of each trace at each location is divided by the sum of all trace values at that location. "percent" is the same but multiplied by 100 to show percentages. If there are multiple subplots, or multiple stackgroups on one subplot, each will be normalized within its own set.

pub fn stack_group(self, stack_group: &str) -> Box<Self>

Set several scatter traces (on the same subplot) to the same stackgroup in order to add their y values (or their x values if orientation is "h"). If blank or omitted this trace will not be stacked. Stacking also turns fill on by default, using "tonexty" ("tonextx") if orientation is "h" ("v") and sets the default mode to "lines" irrespective of point count. You can only stack on a numeric (linear or log) axis. Traces in a stackgroup will only fill to (or be filled to) other traces in the same group. With multiple stackgroups or some traces stacked and some not, if fill-linked traces are not already consecutive, the later ones will be pushed down in the drawing order.

pub fn marker(self, marker: Marker) -> Box<Self>

Determines how points are displayed and joined.

pub fn line(self, line: Line) -> Box<Self>

Line display properties.

pub fn text_font(self, text_font: Font) -> Box<Self>

Sets the text font.

pub fn error_x(self, error_x: ErrorData) -> Box<Self>

x-axis error display properties.

pub fn error_y(self, error_y: ErrorData) -> Box<Self>

y-axis error display properties.

pub fn clip_on_axis(self, clip_on_axis: bool) -> Box<Self>

Determines whether or not markers and text nodes are clipped about the subplot axes. To show markers and text nodes above axis lines and tick labels, make sure to set xaxis.layer and yaxis.layer to "below traces".

pub fn connect_gaps(self, connect_gaps: bool) -> Box<Self>

Determines whether or not gaps (i.e. {nan} or missing values) in the provided data arrays are connected.

pub fn fill(self, fill: Fill) -> Box<Self>

Sets the area to fill with a solid color. Defaults to "none" unless this trace is stacked, then it gets "tonexty" ("tonextx") if orientation is "v" ("h") Use with fillcolor if not "none". "tozerox" and "tozeroy" fill to x=0 and y=0 respectively. "tonextx" and "tonexty" fill between the endpoints of this trace and the endpoints of the trace before it, connecting those endpoints with straight lines (to make a stacked area graph); if there is no trace before it, they behave like "tozerox" and "tozeroy". "toself" connects the endpoints of the trace (or each segment of the trace if it has gaps) into a closed shape. "tonext" fills the space between two traces if one completely encloses the other (eg consecutive contour lines), and behaves like "toself" if there is no trace before it. "tonext" should not be used if one trace does not enclose the other. Traces in a stackgroup will only fill to (or be filled to) other traces in the same group. With multiple stackgroups or some traces stacked and some not, if fill-linked traces are not already consecutive, the later ones will be pushed down in the drawing order.

pub fn fill_color<C: Color>(self, fill_color: C) -> Box<Self>

Sets the fill color. Defaults to a half-transparent variant of the line color, marker color, or marker line color, whichever is available.

pub fn hover_label(self, hover_label: Label) -> Box<Self>

Properties of label displayed on mouse hover.

pub fn hover_on(self, hover_on: &str) -> Box<Self>

Do the hover effects highlight individual points (markers or line points) or do they highlight filled regions? If the fill is "toself" or "tonext" and there are no markers or text, then the default is "fills", otherwise it is "points".

pub fn stack_gaps(self, stack_gaps: &str) -> Box<Self>

Only relevant when stack_group is used, and only the first stack_gaps found in the stackgroup will be used - including if visible is set to Visible::LegendOnly but not if it is set to Visible::False. Determines how we handle locations at which other traces in this group have data but this one does not. With "infer zero" we insert a zero at these locations. With "interpolate" we linearly interpolate between existing values, and extrapolate a constant beyond the existing values.

pub fn x_calendar(self, x_calendar: Calendar) -> Box<Self>

Sets the calendar system to use with x date data.

pub fn y_calendar(self, y_calendar: Calendar) -> Box<Self>

Sets the calendar system to use with y date data.